+++ CHANGELOG.md
@@ -11,9 +11,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - On a typical system during daylight hours, the Envoy will execute the task within 15 minutes.
 - More info about Power Production task here: https://support.enphase.com/s/article/How-do-I-disable-and-enable-power-production
 
+## [7.4.0] - (25.07.2023)
+## Changes
+- added *Energy State* contact sensor for production monitoring, which can be used for notification and automations in HomeKit.
+- added *Energy Level* contact sensor for production monitoring, which can be used for notification and automations in HomeKit.
+- config schema updated
+- cleanup
+
 ## [7.3.0] - (20.07.2023)
 ## Changes
-- added *Power Production On/Off* contact sensors for production monitoring, which can be used for notification and automations in HomeKit.
+- added *Power Production On/Off* contact sensor for production monitoring, which can be used for notification and automations in HomeKit.
 - Use encodeURIComponent in EnvoyToken URLs - thanks @chrisjshull 
 - config schema updated
 - cleanup